change pre_dataset to whisper_asr.py
